pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vocli"
description = "Local voice layer for AI coding tools — 100% offline voice assistant for Claude Code"
requires-python = ">=3.10"
dynamic = ["version"]
license = "MIT"
readme = "README.md"
authors = [
    { name = "Shubham" },
]
keywords = ["voice", "mcp", "claude-code", "speech-to-text", "text-to-speech", "local", "privacy"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]
dependencies = [
    "fastmcp>=2.3.2,<3",
    "httpx",
    "sounddevice",
    "numpy",
    "scipy",
    "webrtcvad-wheels>=2.0.14",
    "click>=8.0",
]

[project.urls]
Homepage = "https://github.com/shubham-lohar/vocli"
Repository = "https://github.com/shubham-lohar/vocli"
Issues = "https://github.com/shubham-lohar/vocli/issues"

[project.scripts]
vocli = "vocli.cli:main"

[tool.hatch.version]
path = "vocli/__version__.py"

[tool.hatch.build.targets.wheel]
packages = ["vocli"]
include = [
    "vocli/**/*.py",
    "vocli/data/**/*.json",
]
